feat: refactor part deletion into wheel assignment
